shader: Split translator to multiple files
