+++ src/emulator/shader/src/translator/branch_cond.cpp
@@ -1,0 +1,108 @@
+// Vita3K emulator project
+// Copyright (C) 2019 Vita3K team
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#include <shader/usse_translator.h>
+#include <shader/usse_decoder_helpers.h>
+#include <shader/usse_disasm.h>
+#include <shader/usse_types.h>
+#include <util/log.h>
+
+using namespace shader;
+using namespace usse;
+
+bool USSETranslatorVisitor::vtst(
+    Imm3 pred,
+    Imm1 skipinv,
+    Imm1 onceonly,
+    Imm1 syncstart,
+    Imm1 dest_ext,
+    Imm1 test_flag_2,
+    Imm1 src1_ext,
+    Imm1 src2_ext,
+    Imm1 prec,
+    RepeatCount rpt_count,
+    Imm2 sign_test,
+    Imm2 zero_test,
+    Imm1 test_crcomb_and,
+    Imm3 chan_cc,
+    Imm2 pdst_n,
+    Imm2 dest_bank,
+    Imm2 src1_bank,
+    Imm2 src2_bank,
+    Imm7 dest_n,
+    Imm1 test_wben,
+    Imm2 alu_sel,
+    Imm4 alu_op,
+    Imm7 src1_n,
+    Imm7 src2_n)
+{
+    return true;
+}
+
+bool USSETranslatorVisitor::vtstmsk(
+    Imm3 pred,
+    Imm1 skipinv,
+    Imm1 onceonly,
+    Imm1 syncstart,
+    Imm1 dest_ext,
+    Imm1 test_flag_2,
+    Imm1 src1_ext,
+    Imm1 src2_ext,
+    Imm1 prec,
+    RepeatCount rpt_count,
+    Imm2 sign_test,
+    Imm2 zero_test,
+    Imm1 test_crcomb_and,
+    Imm2 tst_mask_type,
+    Imm2 dest_bank,
+    Imm2 src1_bank,
+    Imm2 src2_bank,
+    Imm7 dest_n,
+    Imm1 test_wben,
+    Imm2 alu_sel,
+    Imm4 alu_op,
+    Imm7 src1_n,
+    Imm7 src2_n)
+{
+    return true;
+}
+
+bool USSETranslatorVisitor::br(
+    ExtPredicate pred,
+    Imm1 syncend,
+    bool exception,
+    bool pwait,
+    Imm1 sync_ext,
+    bool nosched,
+    bool br_monitor,
+    bool save_link,
+    Imm1 br_type,
+    Imm1 any_inst,
+    Imm1 all_inst,
+    std::uint32_t br_off)
+{
+    Opcode op = (br_type == 0) ? Opcode::BA : Opcode::BR;
+
+    if (op == Opcode::BR && (br_off & (1 << 19))) {
+        // PC bits on SGX543 is 20 bits
+        br_off |= 0xFFFFFFFF << 20;
+    }
+
+    LOG_DISASM("{:016x}: {}{} #{}", m_instr, disasm::e_predicate_str(pred), (br_type == 0) ? "BA" : "BR", br_off);
+
+    return true;
+}
