+++ src/emulator/shader/src/translator/data.cpp
@@ -1,0 +1,276 @@
+// Vita3K emulator project
+// Copyright (C) 2019 Vita3K team
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#include <shader/usse_translator.h>
+
+#include <SPIRV/GLSL.std.450.h>
+#include <SPIRV/SpvBuilder.h>
+
+#include <shader/usse_decoder_helpers.h>
+#include <shader/usse_disasm.h>
+#include <shader/usse_types.h>
+#include <util/log.h>
+
+using namespace shader;
+using namespace usse;
+
+bool USSETranslatorVisitor::vmov(
+    ExtPredicate pred,
+    bool skipinv,
+    Imm1 test_bit_2,
+    Imm1 src2_bank_sel,
+    bool syncstart,
+    Imm1 dest_bank_ext,
+    Imm1 end_or_src0_bank_ext,
+    Imm1 src1_bank_ext,
+    Imm1 src2_bank_ext,
+    MoveType move_type,
+    RepeatCount repeat_count,
+    bool nosched,
+    DataType move_data_type,
+    Imm1 test_bit_1,
+    Imm4 src0_swiz,
+    Imm1 src0_bank_sel,
+    Imm2 dest_bank_sel,
+    Imm2 src1_bank_sel,
+    Imm2 src0_comp_sel,
+    Imm4 dest_mask,
+    Imm6 dest_n,
+    Imm6 src0_n,
+    Imm6 src1_n,
+    Imm6 src2_n) {
+    Instruction inst{};
+
+    static const Opcode tb_decode_vmov[] = {
+        Opcode::VMOV,
+        Opcode::VMOVC,
+        Opcode::VMOVCU8,
+        Opcode::INVALID,
+    };
+
+    inst.opcode = tb_decode_vmov[(Imm3)move_type];
+    dest_mask = decode_write_mask(dest_mask, move_data_type == DataType::F16);
+
+    // TODO: dest mask
+    // TODO: flags
+    // TODO: test type
+
+    const bool is_double_regs = move_data_type == DataType::C10 || move_data_type == DataType::F16 || move_data_type == DataType::F32;
+    const bool is_conditional = (move_type != MoveType::UNCONDITIONAL);
+
+    // Decode operands
+    uint8_t reg_bits = is_double_regs ? 7 : 6;
+
+    inst.opr.dest = decode_dest(dest_n, dest_bank_sel, dest_bank_ext, is_double_regs, reg_bits, m_second_program);
+    inst.opr.src1 = decode_src12(src1_n, src1_bank_sel, src1_bank_ext, is_double_regs, reg_bits, m_second_program);
+
+    // Velocity uses a vec4 table, non-extended, so i assumes type=vec4, extended=false
+    inst.opr.src1.swizzle = decode_vec34_swizzle(src0_swiz, false, 2);
+
+    inst.opr.src1.type = move_data_type;
+    inst.opr.dest.type = move_data_type;
+
+    // TODO: adjust dest mask if needed
+
+    if (is_conditional) {
+        LOG_WARN("Conditional vmov instructions unsupported");
+        return false;
+    }
+
+    if (inst.opr.dest.bank == RegisterBank::SPECIAL || inst.opr.src0.bank == RegisterBank::SPECIAL || inst.opr.src1.bank == RegisterBank::SPECIAL || inst.opr.src2.bank == RegisterBank::SPECIAL) {
+        LOG_WARN("Special regs unsupported");
+        return false;
+    }
+
+    std::string disasm_str = fmt::format("{:016x}: {}{}.{} {} {}", m_instr, disasm::e_predicate_str(pred), disasm::opcode_str(inst.opcode), disasm::data_type_str(move_data_type),
+        disasm::operand_to_str(inst.opr.dest, dest_mask), disasm::operand_to_str(inst.opr.src1, dest_mask));
+
+    // if (is_conditional) {
+    //     inst.operands.src0 = decode_src0(src0_n, src0_bank_sel, end_or_src0_bank_ext, is_double_regs);
+    //     inst.operands.src2 = decode_src12(src2_n, src2_bank_sel, src2_bank_ext, is_double_regs);
+    // }
+
+    disasm_str += fmt::format(" {} {}", disasm::operand_to_str(inst.opr.dest, dest_mask), disasm::operand_to_str(inst.opr.src1, dest_mask));
+    LOG_DISASM(disasm_str);
+
+    // Recompile
+
+    m_b.setLine(usse::instr_idx);
+
+    BEGIN_REPEAT(repeat_count, 2)
+    spv::Id source = load(inst.opr.src1, dest_mask, repeat_offset);
+    store(inst.opr.dest, source, dest_mask, repeat_offset);
+    END_REPEAT()
+
+    return true;
+}
+
+bool USSETranslatorVisitor::vpck(
+    ExtPredicate pred,
+    bool skipinv,
+    bool nosched,
+    Imm1 src2_bank_sel,
+    bool syncstart,
+    Imm1 dest_bank_ext,
+    Imm1 end,
+    Imm1 src1_bank_ext,
+    Imm1 src2_bank_ext,
+    RepeatCount repeat_count,
+    Imm3 src_fmt,
+    Imm3 dest_fmt,
+    Imm4 dest_mask,
+    Imm2 dest_bank_sel,
+    Imm2 src1_bank_sel,
+    Imm7 dest_n,
+    Imm2 comp_sel_3,
+    Imm1 scale,
+    Imm2 comp_sel_1,
+    Imm2 comp_sel_2,
+    Imm5 src1_n,
+    Imm1 comp0_sel_bit1,
+    Imm4 src2_n,
+    Imm1 comp_sel_0_bit0) {
+    Instruction inst{};
+
+    // TODO: There are some combinations that are invalid.
+    const DataType dest_data_type_table[] = {
+        DataType::UINT8,
+        DataType::INT8,
+        DataType::O8,
+        DataType::UINT16,
+        DataType::INT16,
+        DataType::F16,
+        DataType::F32,
+        DataType::C10
+    };
+
+    const DataType src_data_type_table[] = {
+        DataType::UINT8,
+        DataType::INT8,
+        DataType::O8,
+        DataType::UINT16,
+        DataType::INT16,
+        DataType::F16,
+        DataType::F32,
+        DataType::C10
+    };
+
+    const Opcode op_table[][static_cast<int>(DataType::TOTAL_TYPE)] = {
+        { Opcode::VPCKU8U8,
+            Opcode::VPCKU8S8,
+            Opcode::VPCKU8O8,
+            Opcode::VPCKU8U16,
+            Opcode::VPCKU8S16,
+            Opcode::VPCKU8F16,
+            Opcode::VPCKU8F32,
+            Opcode::INVALID },
+        { Opcode::VPCKS8U8,
+            Opcode::VPCKS8S8,
+            Opcode::VPCKS8O8,
+            Opcode::VPCKS8U16,
+            Opcode::VPCKS8S16,
+            Opcode::VPCKS8F16,
+            Opcode::VPCKS8F32,
+            Opcode::INVALID },
+        { Opcode::VPCKO8U8,
+            Opcode::VPCKO8S8,
+            Opcode::VPCKO8O8,
+            Opcode::VPCKO8U16,
+            Opcode::VPCKO8S16,
+            Opcode::VPCKO8F16,
+            Opcode::VPCKO8F32,
+            Opcode::INVALID },
+        { Opcode::VPCKU16U8,
+            Opcode::VPCKU16S8,
+            Opcode::VPCKU16O8,
+            Opcode::VPCKU16U16,
+            Opcode::VPCKU16S16,
+            Opcode::VPCKU16F16,
+            Opcode::VPCKU16F32,
+            Opcode::INVALID },
+        { Opcode::VPCKS16U8,
+            Opcode::VPCKS16S8,
+            Opcode::VPCKS16O8,
+            Opcode::VPCKS16U16,
+            Opcode::VPCKS16S16,
+            Opcode::VPCKS16F16,
+            Opcode::VPCKS16F32,
+            Opcode::INVALID },
+        { Opcode::VPCKF16U8,
+            Opcode::VPCKF16S8,
+            Opcode::VPCKF16O8,
+            Opcode::VPCKF16U16,
+            Opcode::VPCKF16S16,
+            Opcode::VPCKF16F16,
+            Opcode::VPCKF16F32,
+            Opcode::VPCKF16C10 },
+        { Opcode::VPCKF32U8,
+            Opcode::VPCKF32S8,
+            Opcode::VPCKF32O8,
+            Opcode::VPCKF32U16,
+            Opcode::VPCKF32S16,
+            Opcode::VPCKF32F16,
+            Opcode::VPCKF32F32,
+            Opcode::VPCKF32C10 },
+        { Opcode::INVALID,
+            Opcode::INVALID,
+            Opcode::INVALID,
+            Opcode::INVALID,
+            Opcode::INVALID,
+            Opcode::VPCKC10F16,
+            Opcode::VPCKC10F32,
+            Opcode::VPCKC10C10 }
+    };
+
+    inst.opcode = op_table[dest_fmt][src_fmt];
+
+    std::string disasm_str = fmt::format("{:016x}: {}{}", m_instr, disasm::e_predicate_str(pred), disasm::opcode_str(inst.opcode));
+
+    inst.opr.dest = decode_dest(dest_n, dest_bank_sel, dest_bank_ext, false, 7, m_second_program);
+    inst.opr.src1 = decode_src12(src1_n, src1_bank_sel, src1_bank_ext, true, 7, m_second_program);
+
+    inst.opr.dest.type = dest_data_type_table[dest_fmt];
+    inst.opr.src1.type = src_data_type_table[src_fmt];
+
+    if (inst.opr.dest.bank == RegisterBank::SPECIAL || inst.opr.src0.bank == RegisterBank::SPECIAL || inst.opr.src1.bank == RegisterBank::SPECIAL || inst.opr.src2.bank == RegisterBank::SPECIAL) {
+        LOG_WARN("Special regs unsupported");
+        return false;
+    }
+
+    Imm2 comp_sel_0 = comp_sel_0_bit0;
+
+    if (inst.opr.src1.type == DataType::F32)
+        comp_sel_0 |= (comp0_sel_bit1 & 1) << 1;
+    else
+        comp_sel_0 |= (src2_n & 1) << 1;
+
+    inst.opr.src1.swizzle = SWIZZLE_CHANNEL_4_CAST(comp_sel_0, comp_sel_1, comp_sel_2, comp_sel_3);
+
+    disasm_str += fmt::format(" {} {}", disasm::operand_to_str(inst.opr.dest, dest_mask), disasm::operand_to_str(inst.opr.src1, dest_mask));
+    LOG_DISASM(disasm_str);
+
+    // Recompile
+
+    m_b.setLine(usse::instr_idx);
+
+    BEGIN_REPEAT(repeat_count, 2)
+    spv::Id source = load(inst.opr.src1, dest_mask, repeat_offset);
+    store(inst.opr.dest, source, dest_mask, repeat_offset);
+    END_REPEAT()
+
+    return true;
+}
