+++ src/emulator/shader/src/translator/special.cpp
@@ -1,0 +1,48 @@
+// Vita3K emulator project
+// Copyright (C) 2019 Vita3K team
+//
+// This program is free software; you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation; either version 2 of the License, or
+// (at your option) any later version.
+//
+// This program is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License along
+// with this program; if not, write to the Free Software Foundation, Inc.,
+// 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
+
+#include <shader/usse_translator.h>
+
+#include <SPIRV/GLSL.std.450.h>
+#include <SPIRV/SpvBuilder.h>
+
+#include <shader/usse_decoder_helpers.h>
+#include <shader/usse_disasm.h>
+#include <shader/usse_types.h>
+#include <util/log.h>
+
+using namespace shader;
+using namespace usse;
+
+bool USSETranslatorVisitor::phas() {
+    usse::instr_idx--;
+    LOG_DISASM("{:016x}: PHAS", m_instr);
+    return true;
+}
+
+bool USSETranslatorVisitor::nop() {
+    LOG_DISASM("{:016x}: NOP", m_instr);
+    return true;
+}
+
+bool USSETranslatorVisitor::spec(
+    bool special,
+    SpecialCategory category) {
+    usse::instr_idx--;
+    LOG_DISASM("{:016x}: SPEC category: {}, special: {}", m_instr, (uint8_t)category, special);
+    return true;
+}
